doc update

+++ silvia_arduino/silvia_timer.cpp
@@ -2,7 +2,7 @@
 
 //unsigned long timer_start;
 unsigned long start_time;
-int millis_passed;
+unsigned long millis_passed;
 
 void timerReset() {
   millis_passed = 0;
@@ -12,7 +12,7 @@ void timerStart() {
   start_time = millis();
 }
 
-int timerUpdate() {
+unsigned long timerUpdate() {
   if (brew_output.getStatus()) {
     millis_passed = millis() - start_time;
   }
